+++ zxcvbn-core/Scoring/DateGuessesCalculator.cs
@@ -4,11 +4,22 @@ using Zxcvbn.Matcher.Matches;
 
 namespace Zxcvbn.Scoring
 {
+    /// <summary>
+    /// Estimates the number of attempts needed to guess the date.
+    /// </summary>
     internal class DateGuessesCalculator
     {
+        /// <summary>
+        /// The minimum distance between the reference date and the provided date.
+        /// </summary>
         public const int MinimumYearSpace = 20;
 
-        public static double CalculateGuesses(DateMatch match)
+        /// <summary>
+        /// Estimates the attempts required to guess the password.
+        /// </summary>
+        /// <param name="match">The match.</param>
+        /// <returns>The guesses estimate.</returns>
+        public static long CalculateGuesses(DateMatch match)
         {
             var yearSpace = Math.Max(Math.Abs(match.Year - DateMatcher.ReferenceYear), MinimumYearSpace);
